Split OrderButtonsChromeLogic into smaller classes.
