+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -624,6 +624,7 @@
     <Compile Include="Graphics\TilesetSpecificSpriteSequence.cs" />
     <Compile Include="Traits\Pluggable.cs" />
     <Compile Include="Traits\Plug.cs" />
+    <Compile Include="Widgets\Logic\Ingame\MenuButtonsChromeLogic.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
