+++ mods/cnc/chrome/ingame.yaml
@@ -61,7 +61,7 @@ Container@INGAME_ROOT:
 		TooltipContainer@TOOLTIP_CONTAINER:
 
 Container@OBSERVER_WIDGETS:
-	Logic: OrderButtonsChromeLogic
+	Logic: MenuButtonsChromeLogic
 	Children:
 		MenuButton@OPTIONS_BUTTON:
 			Key: escape
@@ -237,7 +237,7 @@ Container@PLAYER_WIDGETS:
 					ReadyText: Ready
 					HoldText: On Hold
 		Background@SIDEBAR_BACKGROUND:
-			Logic: OrderButtonsChromeLogic
+			Logic: MenuButtonsChromeLogic
 			X: WINDOW_RIGHT - 204
 			Y: 30
 			Width: 194
@@ -260,6 +260,7 @@ Container@PLAYER_WIDGETS:
 							ImageCollection: order-icons
 							ImageName: options
 				Button@BEACON_BUTTON:
+					Logic: BeaconOrderButtonLogic
 					X: 62
 					Y: 0-24
 					Width: 30
@@ -273,6 +274,7 @@ Container@PLAYER_WIDGETS:
 							Y: 5
 							ImageCollection: order-icons
 				Button@SELL_BUTTON:
+					Logic: SellOrderButtonLogic
 					X: 102
 					Y: 0-24
 					Width: 30
@@ -286,6 +288,7 @@ Container@PLAYER_WIDGETS:
 							Y: 5
 							ImageCollection: order-icons
 				Button@REPAIR_BUTTON:
+					Logic: RepairOrderButtonLogic
 					X: 142
 					Y: 0-24
 					Width: 30
