+++ mods/d2k/chrome/ingame-observer.yaml
@@ -1,16 +1,13 @@
 Container@OBSERVER_WIDGETS:
+	Logic: MenuButtonsChromeLogic
 	Children:
 		MenuButton@OPTIONS_BUTTON:
-			Logic: OrderButtonsChromeLogic
-			X: 0
-			Y: 0
 			Width: 160
 			Height: 25
 			Text: Options (Esc)
 			Font: Bold
 			Key: escape
 		MenuButton@OBSERVER_STATS_BUTTON:
-			Logic: OrderButtonsChromeLogic
 			MenuContainer: INGAME_OBSERVERSTATS_BG
 			HideIngameUI: False
 			Pause: False
