+++ mods/ra/chrome/ingame-player.yaml
@@ -36,12 +36,12 @@ Container@PLAYER_WIDGETS:
 			ClickThrough: false
 			Children:
 				Container@TOP_BUTTONS:
-					Logic: OrderButtonsChromeLogic
+					Logic: MenuButtonsChromeLogic
 					X: 9
 					Y: 7
 					Children:
 						Button@BEACON_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: BeaconOrderButtonLogic, AddRaceSuffixLogic
 							Width: 28
 							Height: 28
 							Background: sidebar-button
@@ -54,7 +54,7 @@ Container@PLAYER_WIDGETS:
 									Y: 6
 									ImageCollection: order-icons
 						Button@SELL_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: SellOrderButtonLogic, AddRaceSuffixLogic
 							X: 32
 							Width: 28
 							Height: 28
@@ -68,7 +68,7 @@ Container@PLAYER_WIDGETS:
 									Y: 6
 									ImageCollection: order-icons
 						Button@POWER_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: PowerdownOrderButtonLogic, AddRaceSuffixLogic
 							X: 64
 							Width: 28
 							Height: 28
@@ -82,7 +82,7 @@ Container@PLAYER_WIDGETS:
 									Y: 6
 									ImageCollection: order-icons
 						Button@REPAIR_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: RepairOrderButtonLogic, AddRaceSuffixLogic
 							X: 96
 							Width: 28
 							Height: 28
