+++ mods/ts/chrome/ingame-player.yaml
@@ -36,7 +36,7 @@ Container@PLAYER_WIDGETS:
 			ClickThrough: false
 			Children:
 				Container@TOP_BUTTONS:
-					Logic: OrderButtonsChromeLogic
+					Logic: MenuButtonsChromeLogic
 					X: 0
 					Y: 21
 					Children:
@@ -58,7 +58,7 @@ Container@PLAYER_WIDGETS:
 									ImageCollection: order-icons
 									ImageName: debug
 						Button@REPAIR_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: RepairOrderButtonLogic, AddRaceSuffixLogic
 							X: 43
 							Width: 30
 							Height: 31
@@ -73,7 +73,7 @@ Container@PLAYER_WIDGETS:
 									Y: 0
 									ImageCollection: order-icons
 						Button@SELL_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: SellOrderButtonLogic, AddRaceSuffixLogic
 							X: 73
 							Width: 30
 							Height: 31
@@ -88,7 +88,7 @@ Container@PLAYER_WIDGETS:
 									Y: 0
 									ImageCollection: order-icons
 						Button@BEACON_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: BeaconOrderButtonLogic, AddRaceSuffixLogic
 							X: 103
 							Width: 30
 							Height: 31
@@ -103,7 +103,7 @@ Container@PLAYER_WIDGETS:
 									Y: 0
 									ImageCollection: order-icons
 						Button@POWER_BUTTON:
-							Logic: AddRaceSuffixLogic
+							Logic: PowerdownOrderButtonLogic, AddRaceSuffixLogic
 							X: 133
 							Width: 30
 							Height: 31
